+++ 01 - BruteForce/Combinatorica/header/Permutazioni.h
@@ -3,6 +3,6 @@
 
 #include <stddef.h>
 
-void risposte(size_t *a, size_t j, size_t k, size_t length);
+void risposte(int *a, int j, int k, int length);
 
 #endif // PERMUTAZIONI_H
