+++ 01 - BruteForce/Combinatorica/header/PermutazioniRpt.h
@@ -3,6 +3,6 @@
 
 #include <stddef.h>
 
-void risposte(size_t *a, size_t *soluzioni, size_t j, size_t k, size_t length);
+void risposte(int *a, int *soluzioni, int j, int k, int length);
 
 #endif // PERMUTAZIONIRPT_H
