+++ 01 - BruteForce/Combinatorica/header/Sottoinsiemi.h
@@ -1,12 +1,12 @@
 #ifndef SOTTOINSIEMI_H
 #define SOTTOINSIEMI_H
 
-#include <stddef.h>
 #include <stdbool.h>
+#include <stddef.h>
 
-void risposta(size_t *a, bool *r, size_t j, size_t length);
-void sottoinsiemi(size_t *a, size_t length);
-void toString(size_t *a, bool *r, size_t length);
-void risposte(size_t *a, size_t *j, size_t k, size_t length);
+void risposta(int *a, bool *r, int j, int length);
+void sottoinsiemi(int *a, int length);
+void toString(int *a, bool *r, int length);
+void risposte(int *a, int *j, int k, int length);
 
 #endif // SOTTOINSIEMI_H
