+++ 01 - BruteForce/Combinatorica/header/common.h
@@ -5,7 +5,7 @@
 
 void swap(size_t *a, size_t i, size_t j);
 
-void print_array(size_t *array, size_t length);
+void print_array(int *array, int length);
 
 char *toString(size_t *a, int k);
